Created the mockStorage.js file.
