+++ src/modules/interactive.test.js
@@ -1,42 +1,14 @@
 const { toggleTaskComplete, removeAllCompleted } = require('./interactive');
-
-// This is a simple mockStorage object to simulate localStorage behavior
-const mockStorage = {
-  data: {}, // This object will store data as if it were in localStorage
-
-  getItem: function (key) {
-    return this.data[key] || null;
-  },
-
-  setItem: function (key, value) {
-    this.data[key] = value;
-  },
-};
+const { mockStorage } = require('./mockStorage');
 
 /* John: toggleTaskComplete() */
 describe('Toggle Task Functionality', () => {
   test('Should change the status to Checked', () => {
-    const collection = [
-      {
-        index: 1,
-        description: 'Task 1',
-        complete: false,
-      },
-      {
-        index: 2,
-        description: 'Task 2',
-        complete: true,
-      },
-      {
-        index: 3,
-        description: 'Task 3',
-        complete: false,
-      },
-    ];
-
+    const collection = JSON.parse(mockStorage.getItem('collection'));
     toggleTaskComplete(0, collection);
     toggleTaskComplete(1, collection);
     toggleTaskComplete(2, collection);
+    toggleTaskComplete(3, collection);
 
     // Update localStorage after calling toggleTaskComplete()
     mockStorage.setItem('collection', JSON.stringify(collection));
@@ -47,6 +19,7 @@ describe('Toggle Task Functionality', () => {
       { index: 1, description: 'Task 1', complete: true },
       { index: 2, description: 'Task 2', complete: false },
       { index: 3, description: 'Task 3', complete: true },
+      { index: 4, description: 'Task 4', complete: false },
     ]);
   });
 });
@@ -59,6 +32,7 @@ describe('Remove All Completed Functionality', () => {
 
     expect(collection).toEqual([
       { index: 1, description: 'Task 2', complete: false },
+      { index: 2, description: 'Task 4', complete: false },
     ]);
   });
 });
